requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.2
pydantic-settings==2.3.4
python-multipart==0.0.6

# ML and AI
google-genai==0.8.0
openai>=1.40.0
pillow==10.1.0
numpy==1.24.3

# Database and Persistence
sqlalchemy==2.0.23
alembic==1.13.0

# Observability and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# HTTP and Utilities
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0
tenacity==8.2.3

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
ruff==0.1.6
mypy==1.7.1
isort==5.12.0

# Performance Testing
locust==2.17.0

# Data Processing
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
tabulate==0.9.0

# Async and Concurrency
asyncio
aiofiles==23.2.1
aiohttp==3.9.1